requirements.txt
mediapipe==0.9.1.0
opencv-python==4.7.0.68
Pillow==9.1.0
torch==1.13.1
torchvision==0.14.1

